Reimplement how programs are built
There is now a Build<Cpu, Ram> trait and Builder takes an additional lifetime
parameter so that the code written to the tape can borrow from the code
itself. The code is always passed to Machine::program from behind a type
with a stable deref implementation, which means that the code can then
be stored with the tape and the final program be moved at will.

Note that Env is now gone, replaced by Ram. I was very tempted to go Ram/Rem
instead of Ram/Rom but I'm not that obnoxious.